fix binary letters in name will be shown as \ufffd
If you lookup a domain with control letters or extend latin letters, they will be shown as �(\uFFFD).
A vaild domain name shouldn't contains binary letters, but if your device or software doesn't convert i18n names to punycode, it may happen.
For example, if you use `nslookup` to lookup `nic.谷歌` on a device running Windows with Simplified Chinese language, it'll be encoded as GBK, so it'll be `nic.<b9><c8><b8><e8>` instead of `nic.xn--flw351e`.
It should be fine to display them as �(\uFFFD), however if we want to re-encode them to binary, as they're changed to �(\uFFFD), we cannot get the original letters.