refactor rule parser to support custom fields
Now you can use custom field to write your parser instead of only `server` and `file` field, but your parser maybe needs to change some line to make it works.
First you need to set a `init()` method in your parser, and `parse()` method isn't need anymore. Instead of receiving rule content in `parse()` only, you'll receive all the config as an object in `init()`.
And for the compatibility in the future, you should also accept and save the server config, and return it as `{ server }` in `test()` if a domain matches the rule, or `null` if not, instead of just a boolean `true` or `false`.
Then if you want to use some other fields as `server` or `file`, and want DNSProxy to resolve them automaticly, you can set `fieldType` property.